+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add tests for internal schema migrations
 * Speed up `writeMany` by defaulting to use a single statement for all inserts (as opposed to a single transactions with many individual inserts)
   * One can revert to the previous behavior by setting `singleStatement: false` in the call
+* Add `delete(where: )` to delete all rows matching a specific index-query
 
 ## 2.0.0